Added ./src & ./var. Make requirements more versatile
we isolate code source into ./src in order to include it in the Docker
image, and only it. Not the files at the root of the project.

DB file moves to ./var (settings.py updated accordingly)

In the requirements.txt file, we use django<1.9 instead of 1.8.4, which
quickly gets deprecated. This way, the next build of the image will
update the minor version of Django